+++ src/GameLogic/Rules/CaptureZoneRules.cs
@@ -30,6 +30,7 @@ public class CaptureZoneRules : IBattleRules
     public bool UseShrinkingZone => false;
     public bool UsesPowerups => false;
     public int MinPlayersToStart => Constants.MinPlayersToStart;
+    public uint FireCooldownTicks => 10;
 
     public void Initialize(GameRoomState state)
     {