+++ src/GameLogic/Rules/TeamsRules.cs
@@ -24,6 +24,7 @@ public class TeamsRules : IBattleRules
     public bool UseShrinkingZone => false;
     public bool UsesPowerups => true;
     public int MinPlayersToStart => Constants.MinPlayersToStart;
+    public uint FireCooldownTicks => 10;
 
     public void Initialize(GameRoomState state) { }